move dish and category to menu
