--- lib/dish.rb
@@ -1,7 +1,0 @@
-class Dish
-  attr_reader :name, :price
-  def initialize(name, price)
-    @name = name
-    @price = price
-  end
-end
